fix: refactored file tree refactored sign up flow
