+++ src/client/components/index.js
@@ -1,0 +1,9 @@
+//--------- SIGNUP ------------//
+export { default as SignUp } from './auth/SignUp/SignUp';
+export { default as GeneralInfo } from './auth/SignUp/GeneralInfo';
+export { default as Location } from './auth/SignUp/Location';
+export { default as Preferences } from './auth/SignUp/Preferences';
+export { default as Assets } from './auth/SignUp/Assets';
+export { default as Terms } from './auth/SignUp/Terms';
+//--------- DASHBOARD ------------//
+export { default as Try } from './dash/Try';
